refactor(cooklang): extract report widget types, add inline template support
